hard code bindings for tg functions we use
