+++ c/sedona-tg/src/tg_bindgen.rs
@@ -16,7 +16,86 @@
 // under the License.
 #![allow(non_upper_case_globals)]
 #![allow(non_camel_case_types)]
-#![allow(non_snake_case)]
-#![allow(dead_code)]
 
-include!(env!("BINDINGS_PATH"));
+use std::os::raw::{c_char, c_int, c_void};
+
+pub type tg_index = c_int;
+
+pub const tg_index_TG_NONE: tg_index = 0;
+pub const tg_index_TG_DEFAULT: tg_index = 1;
+pub const tg_index_TG_NATURAL: tg_index = 2;
+pub const tg_index_TG_YSTRIPES: tg_index = 3;
+
+#[repr(C)]
+#[derive(Debug, Copy, Clone)]
+pub struct tg_geom {
+    _unused: [u8; 0],
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_free(geom: *mut tg_geom);
+}
+
+unsafe extern "C" {
+    pub fn tg_parse_wktn_ix(wkt: *const c_char, len: usize, ix: tg_index) -> *mut tg_geom;
+}
+
+unsafe extern "C" {
+    pub fn tg_parse_wkb_ix(wkb: *const u8, len: usize, ix: tg_index) -> *mut tg_geom;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_error(geom: *const tg_geom) -> *const c_char;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_wkb(geom: *const tg_geom, dst: *mut u8, n: usize) -> usize;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_wkt(geom: *const tg_geom, dst: *mut ::std::os::raw::c_char, n: usize) -> usize;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_memsize(geom: *const tg_geom) -> usize;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_equals(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_intersects(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_disjoint(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_contains(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_within(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_covers(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_coveredby(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_geom_touches(a: *const tg_geom, b: *const tg_geom) -> bool;
+}
+
+unsafe extern "C" {
+    pub fn tg_env_set_allocator(
+        malloc: Option<unsafe extern "C" fn(arg1: usize) -> *mut c_void>,
+        realloc: Option<unsafe extern "C" fn(arg1: *mut c_void, arg2: usize) -> *mut c_void>,
+        free: Option<unsafe extern "C" fn(arg1: *mut c_void)>,
+    );
+}
